add repeat attribute; make code reusable in one place
