+++ app/src/main/res/layout/activity_main.xml
@@ -17,9 +17,10 @@
             android:id="@+id/stpv_2017"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            app:duration="12000"
+            app:duration="2000"
             app:showPainter="true"
             app:text="2017"
+            app:repeat="RESTART"
             app:textInCenter="true"
             app:textSize="60sp"
             android:layout_weight="1"
@@ -29,9 +30,10 @@
             android:id="@+id/stpv_chicken"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            app:duration="12000"
+            app:duration="3000"
+            app:repeat="REVERSE"
             app:showPainter="true"
-            app:text="鸡年"
+            app:text="abc"
             app:textInCenter="true"
             app:textSize="60sp"
             android:layout_weight="1"
@@ -49,9 +51,10 @@
             android:id="@+id/stpv_2018"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            app:duration="12000"
+            app:duration="1000"
             app:showPainter="true"
             app:text="2018"
+            app:repeat="RESTART"
             app:textInCenter="true"
             app:textStrokeColor="@android:color/holo_red_light"
             app:paintStrokeColor="@android:color/holo_red_light"
@@ -79,7 +82,8 @@
         android:id="@+id/atpv_1"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        app:duration="12000"
+        app:duration="3000"
+        app:repeat="RESTART"
         app:showPainter="true"
         app:text="炎之铠"
         app:textStrokeColor="@android:color/holo_orange_light"
@@ -108,7 +112,8 @@
         android:id="@+id/atpv_2"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        app:duration="12000"
+        app:duration="3000"
+        app:repeat="REVERSE"
         app:showPainter="true"
         app:text="新春快乐"
         app:textInCenter="true"
