+++ lib/src/main/java/yanzhikai/textpath/AsyncTextPathView.java
@@ -1,5 +1,6 @@
 package yanzhikai.textpath;
 
+import android.animation.ValueAnimator;
 import android.content.Context;
 import android.graphics.Path;
 import android.support.annotation.Nullable;
@@ -105,24 +106,6 @@ public class AsyncTextPathView extends TextPathView {
         initTextPath();
         clear();
         requestLayout();
-    }
-
-    /**
-     * 开始绘制文字路径动画
-     * @param start 路径比例，范围0-1
-     * @param end 路径比例，范围0-1
-     */
-    public void startAnimation(float start, float end){
-        if (!isProgressValid(start) || !isProgressValid(end)){
-            return;
-        }
-        if (mAnimator != null) {
-            mAnimator.cancel();
-        }
-        initAnimator(start, end);
-        initTextPath();
-        showPainterActually = showPainter;
-        mAnimator.start();
     }
 
     //设置画笔特效
